Start reorganizing impala code into ibis.impala
