+++ ibis/impala/identifiers.py
